Move RedBee extractors to one file
